Refactor Snake Game into hooks, outside of the main file
